Add files via upload
An iteration of part of the self driving algorithm for an autonomous race car. This program will use a StereoLabs ZED2 camera attached to an RC car to perform perception, object tracking, localisation and mapping. The car is driven around the track as part of a "formation lap" to collect information about it's surroundings and it's position in the surroundings over time.